Refactorización y mejoras en acceso a datos y cálculos
Se centralizó el acceso a datos mediante `Dapper` y cadenas de conexión en `App.config`. Se reemplazaron variables `double` por `decimal` para cálculos financieros en `FrmViajes` y `FrmViajesRetro`. Se añadieron repositorios (`RepositorioBase`, `RepositorioAsistencias`, etc.) y modelos (`TipoAsistencia`, `DiaFeriado`, etc.) para mejorar la modularidad.

Se añadió el formulario `frmHistorialSalarios` para gestionar salarios. Se optimizó la lógica de carga de datos en `frmAsistencias` y se eliminaron consultas SQL hardcodeadas. Se actualizó el framework a .NET 4.7.2 y se incluyó la biblioteca `Dapper`. Se eliminaron clases obsoletas como `DbManager`.

+++ ADIGGM/Visores/FrmVisorViajes.cs
@@ -179,11 +179,11 @@ namespace ADIGGM.Formularios_Base.Visores
             int IdTipoVeh = int.Parse(dgvVisorViajes.CurrentRow.Cells["idTipoVeh"].Value.ToString());
             int IdRuta = int.Parse(dgvVisorViajes.CurrentRow.Cells["idRuta"].Value.ToString());
             int IdVehiculo = int.Parse(dgvVisorViajes.CurrentRow.Cells["idVehiculo"].Value.ToString());
-            double Cantidad = double.Parse(dgvVisorViajes.CurrentRow.Cells["cantidad"].Value.ToString());
-            double Tarifa = double.Parse(dgvVisorViajes.CurrentRow.Cells["tarifa"].Value.ToString());
-            double ISV = double.Parse(dgvVisorViajes.CurrentRow.Cells["iSV"].Value.ToString());
-            double SubTotal = double.Parse(dgvVisorViajes.CurrentRow.Cells["subtotal"].Value.ToString());
-            double Total = double.Parse(dgvVisorViajes.CurrentRow.Cells["total"].Value.ToString());
+            decimal Cantidad = decimal.Parse(dgvVisorViajes.CurrentRow.Cells["cantidad"].Value.ToString());
+            decimal Tarifa = decimal.Parse(dgvVisorViajes.CurrentRow.Cells["tarifa"].Value.ToString());
+            decimal ISV = decimal.Parse(dgvVisorViajes.CurrentRow.Cells["iSV"].Value.ToString());
+            decimal SubTotal = decimal.Parse(dgvVisorViajes.CurrentRow.Cells["subtotal"].Value.ToString());
+            decimal Total = decimal.Parse(dgvVisorViajes.CurrentRow.Cells["total"].Value.ToString());
             string Observaciones = dgvVisorViajes.CurrentRow.Cells["observaciones"].Value.ToString();
 
             if(IdTipoVeh == IdTipoVehiculo)
@@ -310,11 +310,11 @@ namespace ADIGGM.Formularios_Base.Visores
             int IdTipoVeh = int.Parse(dgvVisorViajes.CurrentRow.Cells["idTipoVeh"].Value.ToString());
             int IdRuta = int.Parse(dgvVisorViajes.CurrentRow.Cells["idRuta"].Value.ToString());
             int IdVehiculo = int.Parse(dgvVisorViajes.CurrentRow.Cells["idVehiculo"].Value.ToString());
-            double Cantidad = double.Parse(dgvVisorViajes.CurrentRow.Cells["cantidad"].Value.ToString());
-            double Tarifa = double.Parse(dgvVisorViajes.CurrentRow.Cells["tarifa"].Value.ToString());
-            double ISV = double.Parse(dgvVisorViajes.CurrentRow.Cells["iSV"].Value.ToString());
-            double SubTotal = double.Parse(dgvVisorViajes.CurrentRow.Cells["subtotal"].Value.ToString());
-            double Total = double.Parse(dgvVisorViajes.CurrentRow.Cells["total"].Value.ToString());
+            decimal Cantidad = decimal.Parse(dgvVisorViajes.CurrentRow.Cells["cantidad"].Value.ToString());
+            decimal Tarifa = decimal.Parse(dgvVisorViajes.CurrentRow.Cells["tarifa"].Value.ToString());
+            decimal ISV = decimal.Parse(dgvVisorViajes.CurrentRow.Cells["iSV"].Value.ToString());
+            decimal SubTotal = decimal.Parse(dgvVisorViajes.CurrentRow.Cells["subtotal"].Value.ToString());
+            decimal Total = decimal.Parse(dgvVisorViajes.CurrentRow.Cells["total"].Value.ToString());
             string Observaciones = dgvVisorViajes.CurrentRow.Cells["observaciones"].Value.ToString();
 
             if (IdTipoVeh == IdTipoVehiculo)
